# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LLM-as-a-Judge AI Model Evaluation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Model {
  id: number;
  modelName: string;
  modelSize: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CreateModelBody {
  modelName: string;
  modelSize: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateModelBody {
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  modelSize?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type DatasetDomain = (typeof DatasetDomain)[keyof typeof DatasetDomain];

export const DatasetDomain = {
  Medical: "Medical",
  General: "General",
} as const;

export interface Dataset {
  id: number;
  datasetName: string;
  domain: DatasetDomain;
  questionCount: number;
  createdAt: string;
}

export type CreateDatasetBodyDomain =
  (typeof CreateDatasetBodyDomain)[keyof typeof CreateDatasetBodyDomain];

export const CreateDatasetBodyDomain = {
  Medical: "Medical",
  General: "General",
} as const;

export interface CreateDatasetBody {
  datasetName: string;
  domain: CreateDatasetBodyDomain;
}

export type UploadDatasetBodyFormat =
  (typeof UploadDatasetBodyFormat)[keyof typeof UploadDatasetBodyFormat];

export const UploadDatasetBodyFormat = {
  csv: "csv",
  jsonl: "jsonl",
} as const;

export interface UploadDatasetBody {
  datasetId: number;
  /** Raw CSV or JSONL content as a string */
  content: string;
  format: UploadDatasetBodyFormat;
}

export interface UploadDatasetResponse {
  imported: number;
  skipped: number;
  errors: string[];
}

export type QuestionQuestionType =
  (typeof QuestionQuestionType)[keyof typeof QuestionQuestionType];

export const QuestionQuestionType = {
  MCQ: "MCQ",
  OPEN_ENDED: "OPEN_ENDED",
} as const;

/**
 * JSONB: MCQ choices, must_have, nice_to_have, etc.
 */
export type QuestionMetadata = { [key: string]: unknown };

export interface Question {
  id: number;
  datasetId: number;
  questionText: string;
  goldAnswer: string;
  questionType: QuestionQuestionType;
  /** JSONB: MCQ choices, must_have, nice_to_have, etc. */
  metadata?: QuestionMetadata;
  createdAt: string;
}

export type CreateQuestionBodyQuestionType =
  (typeof CreateQuestionBodyQuestionType)[keyof typeof CreateQuestionBodyQuestionType];

export const CreateQuestionBodyQuestionType = {
  MCQ: "MCQ",
  OPEN_ENDED: "OPEN_ENDED",
} as const;

export type CreateQuestionBodyMetadata = { [key: string]: unknown };

export interface CreateQuestionBody {
  datasetId: number;
  questionText: string;
  goldAnswer: string;
  questionType: CreateQuestionBodyQuestionType;
  metadata?: CreateQuestionBodyMetadata;
}

export interface ModelResponse {
  id: number;
  questionId: number;
  modelId: number;
  responseText: string;
  /** @nullable */
  inferenceTimeMs?: number | null;
  createdAt: string;
  /** @nullable */
  questionText?: string | null;
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  goldAnswer?: string | null;
  /** @nullable */
  questionType?: string | null;
}

export interface GenerateResponsesBody {
  modelId: number;
  /** List of question IDs to generate responses for */
  questionIds?: number[];
  /**
   * Alternatively, generate for all questions in a dataset
   * @nullable
   */
  datasetId?: number | null;
}

export interface GenerateResponsesResult {
  generated: number;
  skipped: number;
  errors: string[];
}

export interface JudgeEvaluation {
  id: number;
  responseId: number;
  judgeModelId: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  score: number;
  reasoning: string;
  evaluatedAt: string;
  /** @nullable */
  questionText?: string | null;
  /** @nullable */
  responseText?: string | null;
  /** @nullable */
  modelName?: string | null;
  /** @nullable */
  judgeModelName?: string | null;
}

export interface RunJudgeBody {
  judgeModelId: number;
  /** Specific response IDs to evaluate */
  responseIds?: number[];
  /**
   * Evaluate all responses for a dataset
   * @nullable
   */
  datasetId?: number | null;
  /**
   * Filter by model
   * @nullable
   */
  modelId?: number | null;
}

export interface RunJudgeResult {
  evaluated: number;
  skipped: number;
  errors: string[];
}

export interface AnalyticsSummary {
  totalModels: number;
  totalDatasets: number;
  totalQuestions: number;
  totalResponses: number;
  totalEvaluations: number;
  /** @nullable */
  averageScore?: number | null;
  mcqCount: number;
  openEndedCount: number;
}

export interface ModelComparisonRow {
  modelId: number;
  modelName: string;
  provider: string;
  /** @nullable */
  avgScore?: number | null;
  totalEvaluations: number;
  totalResponses: number;
  /** @nullable */
  avgInferenceMs?: number | null;
}

export interface ScoreDistributionRow {
  modelId: number;
  modelName: string;
  score: number;
  count: number;
}

export interface ResultRow {
  questionId: number;
  questionText: string;
  goldAnswer: string;
  questionType: string;
  datasetName: string;
  responseId: number;
  modelName: string;
  responseText: string;
  /** @nullable */
  inferenceTimeMs?: number | null;
  /** @nullable */
  evaluationId?: number | null;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  reasoning?: string | null;
  /** @nullable */
  judgeModelName?: string | null;
  /** @nullable */
  evaluatedAt?: string | null;
}

export interface SpearmanResult {
  /** @nullable */
  correlation?: number | null;
  /** @nullable */
  pValue?: number | null;
  sampleSize: number;
  interpretation: string;
  /** @nullable */
  mcqAccuracy?: number | null;
}

export interface ApiKeyStatus {
  openai: boolean;
  deepseek: boolean;
  claude: boolean;
}

export interface SaveApiKeysBody {
  /** @nullable */
  openaiKey?: string | null;
  /** @nullable */
  deepseekKey?: string | null;
  /** @nullable */
  claudeKey?: string | null;
}

export type ListQuestionsParams = {
  /**
   * @nullable
   */
  datasetId?: number | null;
  /**
   * @nullable
   */
  questionType?: ListQuestionsQuestionType;
};

export type ListQuestionsQuestionType =
  | (typeof ListQuestionsQuestionType)[keyof typeof ListQuestionsQuestionType]
  | null;

export const ListQuestionsQuestionType = {
  MCQ: "MCQ",
  OPEN_ENDED: "OPEN_ENDED",
} as const;

export type ListResponsesParams = {
  /**
   * @nullable
   */
  questionId?: number | null;
  /**
   * @nullable
   */
  modelId?: number | null;
};

export type ListEvaluationsParams = {
  /**
   * @nullable
   */
  responseId?: number | null;
  /**
   * @nullable
   */
  judgeModelId?: number | null;
};

export type GetResultsParams = {
  /**
   * @nullable
   */
  datasetId?: number | null;
  /**
   * @nullable
   */
  modelId?: number | null;
};