# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LLM-as-a-Judge AI Model Evaluation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /models:
    get:
      operationId: listModels
      tags: [models]
      summary: List all AI models
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
    post:
      operationId: createModel
      tags: [models]
      summary: Register a new AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelBody"
      responses:
        "201":
          description: Model created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
  /models/{id}:
    get:
      operationId: getModel
      tags: [models]
      summary: Get a model by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Model found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Model not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateModel
      tags: [models]
      summary: Update a model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelBody"
      responses:
        "200":
          description: Model updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
    delete:
      operationId: deleteModel
      tags: [models]
      summary: Delete a model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Model deleted

  # Datasets
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Model:
      type: object
      properties:
        id:
          type: integer
        modelName:
          type: string
        modelSize:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - modelName
        - modelSize
        - createdAt
    CreateModelBody:
      type: object
      properties:
        modelName:
          type: string
        modelSize:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - modelName
        - modelSize
    UpdateModelBody:
      type: object
      properties:
        modelName:
          type: ["string", "null"]
        modelSize:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
...